framework/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

einops==0.8.1
gudhi==3.11.0
matplotlib==3.10.8
numpy==1.24.3
opencv-python-headless==4.11.0.86
panel==1.8.9
pillow==12.0.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
thop==0.1.1-2209072238
timm==1.0.22
torch==2.0.1
torchinfo==1.8.0
torchvision==0.15.2
tqdm==4.67.1
